refactor: drought email api test setup
